requirements.txt
flask>=2.2
requests>=2.28
beautifulsoup4>=4.12
python-dotenv>=1.0
supabase>=1.0.0
stripe>=5.0
playwright>=1.30
